Update h5io_browser to 0.2.9, numpy to 2.4.1, pandas to 3.0.0, pyfileindex to 0.1.3 and pysqa to 0.3.4 (#1957)
* Update pyproject.toml

* Update environment.yml

* Update environment-mini.yml

* Update environment-docs.yml

* Update environment.yml

* Update environment-docs.yml

* Update environment-mini.yml

* Update environment.yml

* Update environment.yml

* Update pyproject.toml

* Update environment.yml

* pandas fixes

+++ tests/unit/archiving/test_export.py
@@ -45,7 +45,7 @@ class TestPack(PyironTestCase):
         # and is compared with the return dataframe from export_database
         directory_to_transfer = os.path.basename(self.pr.path[:-1])
         df_exp = export_database(self.pr.job_table()).dropna(axis=1)
-        df_exp["hamversion"] = float(df_exp["hamversion"])
+        df_exp["hamversion"] = df_exp["hamversion"].astype(float)
         self.assertEqual(df_exp["job"].unique()[0], "toy")
         self.assertEqual(df_exp["id"].unique()[0], 0)
